sched/core: fix SCHED_FEAT overflow on 32-bit architectures
hulk inclusion
category: bugfix
bugzilla: https://atomgit.com/openeuler/kernel/issues/8882

(cherry-pick from OLK-6.6 commit 8b4e3033b531)

--------------------------------

SCHED_FEAT uses 1UL for bit shifting, but on ARM32 (and other 32-bit
architectures), unsigned long is 32 bits. With 37 SCHED_FEAT entries,
the bit shift can overflow beyond bit 31, causing incorrect scheduler
feature detection.

Change all 1UL to 1ULL and change sysctl_sched_features type from
unsigned long/int to unsigned long long to ensure 64-bit width on all
architectures.

Fixes: 765cc3a4b224 ("sched/core: Optimize sched_feat() for !CONFIG_SCHED_DEBUG builds")

+++ kernel/sched/core.c
@@ -134,8 +134,8 @@ static void sched_change_qos_group(struct task_struct *tsk, struct task_group *t
  * at compile time and compiler optimization based on features default.
  */
 #define SCHED_FEAT(name, enabled)	\
-	(1UL << __SCHED_FEAT_##name) * enabled |
-const_debug unsigned long sysctl_sched_features =
+	(1ULL << __SCHED_FEAT_##name) * enabled |
+const_debug unsigned long long sysctl_sched_features =
 #include "features.h"
 	0;
 #undef SCHED_FEAT
